{{cookiecutter.project_slug}}/requirements.txt
Django==5.0.3
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-organizations==2.4.1
django-polymorphic==3.1.0
jsonschema==4.21.1
psycopg[binary]==3.1.18
pytz==2024.1 
channels[daphne]==4.0.0
